Cancel HandoverRetry when it's in wrong network state
Race condition occurs when two same handover retries that are scheduled very close(e.g. modem unthrottles an APN which triggers a retry coincides a retry after handover failure) that both of them passed `!dataNetwork.isConnected()` check at `onDataNetworkHandoverRetry`, and sent two EVENT_START_HANDOVER in DN. One of them successfully starts the handover while the other one is ignored because it's already in handover state, but the ignored retry entry is always set as "not_retried", so in future when we attempt to start another handover, it will be disallowed due to the obsolete retry entry "not_retried".

Bug: 235567654
Test: manual verified
Change-Id: Ief834becf215ab843957f5f9ecf70f28273db0e5