Remap no sim and no service call fail causes to valid telephony codes.
Add remappings so that the modem can specifically report no sim and out
of service cases to Telephony.  In no SIM or out of service cases we
normally get ERROR_UNSPECIFIED or NORMAL_CLEARING from the lower levels;
those codes and the default block have code which checks for out of
service and no sim cases specifically.  In the case of an emergency call
that fails to go through, it is very likely that it was placed when there
was no sim or no service in the first place and it is desirable to be able
to report that to upper levels.

Test: atest GsmCdmaConnectionTest
Bug: 238716207
Change-Id: I792247d00e303ad26562cf727c4a82aad19a063c